ScoreWindow not early also for dancers.

+++ game/screen_sing.cc
@@ -329,7 +329,7 @@ void ScreenSing::draw() {
 			}
 		}
 		else if (!m_audio.isPlaying() || (status == Song::FINISHED
-		  && m_audio.getLength() - time <= (m_song->track_map.empty() ? 3.0 : 0.2) )) {
+		  && m_audio.getLength() - time <= (m_song->track_map.empty() && m_song->danceTracks.empty() ? 3.0 : 0.2) )) {
 			// Time to create the score window
 			m_quitTimer.setValue(QUIT_TIMEOUT);
 			m_engine->kill(); // kill the engine thread (to avoid consuming memory)
